\name{init.terra}
\alias{init.terra}
\title{initializes the terra subsystem}
\usage{
  init.terra(includes, libraries)
}
\arguments{
  \item{includes}{is a vector of include directories}

  \item{libraries}{path to libraries to load}
}
\value{
  TRUE upon success, error if it fails
}
\description{
  initializes the terra subsystem
}

