% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/baseterra.R
\name{terraAddGeneralPaths}
\alias{terraAddGeneralPaths}
\title{Updates a search path}
\usage{
terraAddGeneralPaths(modulenames, whatpath)
}
\arguments{
\item{whatpath}{exampls are package.path or package.cpath}

\item{paths}{a vector of paths to load}
}
\value{
throws an error upon failure
##'
}
\description{
Updates a search path
}
\seealso{
\code{\link{terraAddIncludePaths}},\code{\link{terraLinkLibrary}},,\code{\link{terraAddGeneralPaths}}
}

