% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/baseterra.R
\name{terraStr}
\alias{terraStr}
\title{Evaluates a terra string}
\usage{
terraStr(s)
}
\arguments{
\item{s}{the string to evaluate}
}
\value{
1 or 0 upon success
}
\description{
Evaluates a terra string
}
\seealso{
\code{\link{terra}},\code{\link{terraFile}}
}

