% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/baseterra.R
\name{tinit}
\alias{tinit}
\title{initializes the terra subsystem}
\usage{
tinit(rcppflags = processCppFlags(system("R CMD config --cppflags", intern =
  TRUE)), luastate = NULL, options = list(verbose = 0L, debug = 1L))
}
\arguments{
\item{rcppflags}{provide your own value for R CMD config --cppflags (if you do, a character vector)}

\item{options}{if verbose >0 then lots of output and if debug>0 then line numebrs in stack traces}

\item{luatstate}{a preexisting laustate}
}
\value{
TRUE upon success, error if it fails
}
\description{
initializes the terra subsystem
}

