\name{terraAddIncludePaths}
\alias{terraAddIncludePaths}
\title{Adds include paths to the terra system}
\usage{
  terraAddIncludePaths(paths)
}
\arguments{
  \item{paths}{a vector of paths to load}
}
\value{
  throws an error upon failure
}
\description{
  Adds include paths to the terra system
}
\seealso{
  \code{\link{terraAddRequirePaths}},\code{\link{terraLinkLibrary}}
}

