\name{terraFile}
\alias{terraFile}
\title{Loads code from a terra file
Similar to R's source command, this command runs the terra file through terra}
\usage{
terraFile(filename)
}
\arguments{
  \item{filename}{the path to a terra file}
}
\value{
1 or 0 upon success or failure
}
\description{
Loads code from a terra file Similar to R's source command,
this command runs the terra file through terra
}
\author{
Saptarshi Guha
}
\seealso{
\code{\link{terra}},\code{\link{terraStr}}
}

